Refactor to slim down PlayerManagement, introduce single threading into Ollama processing for all AI.
