+++ docs/versions/1.1.5/V1.1.5_ChangeLog.md
@@ -13,6 +13,9 @@
 - Added new Action for Playing a local sound file when configured in the actions of a line handler.  This can be for something like 'Vote To Kicks' and get a sound to play to get your attention.  See the [Config Tab, Line Handlers PlaySoundAction](../../Config_LineHandlers.md#playsoundaction-definition-in-actions)
 
 ## Behavioural Changes
+- Seperated the Group/Moderation evalutation from the AI evaluation, this should help with the performance of the application as well as allow for the AI evaluation to be disabled if you don't want to use it.  The Group/Moderation evaluation will still run even if the AI evaluation is disabled.
+- Combined queueing of ALL AI Evaluation into a single queue to eliminate the possibility of multiple threads trying to access the Ollama API at the same time.  This should help with the errors comming from the Ollama Cloud API.
+- Semi implemented the abillity to get the public avatar IDs for Avatar Unpack events, it logs the Avatar ID to the log file, but does not yet display it in the UI.  This will be implemented in a future release.  It does rely on the VRCDB to have the AV in their database.
 - AI Evaluation / Ollama performance logging; I have seen some backlogs of the AI evalutation to take a long time to process, so I have added logging to the log file to see how long it takes to process each user.  This should help me identify if there is a performance issue with the AI evaluation and if so, what is causing it.  Most of the time we should see 4 second or less processing of user bios.
 - Moderation Reports:
 	- New Table to hold onto all the in game moderaton reports you have submitted to VRChat.  This will allow you to see what you have reported and when, as well as the ability to export the reports to a CSV file for your own records.
@@ -28,11 +31,14 @@
 	- Check GIST button for Group now overides the Checksum check and will always check the GIST file for updates.  This is to allow users to force a recheck of the GIST file if they suspect it has been updated since the last time it was checked or detect deleted groups in the database (Were UpdatedDate < Now()/Cutton Click Time)
 	- Check GIST button for Group now passes the URL of the GIST file Textbox to the processing code, this allows the primary load process as well as the ability to Merge in other GIST files from other sources; as so long as you don't save the page, the Prior URI will still remain in the Registry for next startup of Tailgrab.
 	- Roundtrip to VRChat API for updating the Group's Name and validating it still exists is now cached for 12 hours, this should help with the performance of the GIST processing as well as reduce the number of API calls to VRChat.
+- Refactoring of the various functionals to make the PlayerManagement code more maintainable and easier to add new features in the future.
 
 ## UI Changes
 
 - Iconography Changes:
 	- Tried to get most of the ENUM/Classifications into some sort of iconography to make it easier to see the state of a user at a glance.  This should make it easier to see the state of a user at a glance, and make it easier to see the state of a user at a glance.
+- Button for Secrets pannel to delete the Two Factor Authentication (2FA) secrets from the application.  This will allow you to reset your 2FA if you have lost access to your 2FA device or if you want to reset your 2FA for any other reason.  This will also allow you to reset your 2FA if you have lost access to your 2FA device or if you want to reset your 2FA for any other reason.
+- Button for AI Evaluation pannel to delete the Ollama API Key from the application.  This will allow you to reset your Ollama API Key if you have lost access to your Ollama API Key or if you want to reset your Ollama API Key for any other reason.  This will also allow you to reset your Ollama API Key if you have lost access to your Ollama API Key or if you want to reset your Ollama API Key for any other reason.
 
 **Alert Level Iconography**
